Show multi-category commits under each category
A common kind of category in SerenityOS commits are the multi-category
specifiers, when a commit touches more than one area of the system.
There's a standardized way to specify those with a "+", an example would
be "LibJS+Kernel+AK: ...".

This change breaks up such commit categories, so that commits get
categorized under each individual category. In the above example, the
commit would appear under "LibJS" and "Kernel" and "AK". While this does
duplicate commits and make some categories larger, it removes the weird
single-commit "+" headers on the page, making it easier to read and
navigate.

For achieving this, a new function that extracts categories from commit
titles was created, which is capable of returning more than one
category.