Fetch multiple commit pages at the same time
This drastically improves the speed of the commit fetch.
This works by taking the Link header that the GH API gives us and extracting the "Last" page from it.
We then queue up every single page for fetching and return a single promise with Promise.all.
The downside is that the user of paginate must now flatten the array, but that's a small price to pay.